+++ lib/src/input_types.dart
@@ -13,4 +13,5 @@ enum InputType {
   multipleChoice,
   otp,
   dropdown,
+  dynamicKeyValue
 }